Break webview host handler cycles via shared types (oh-tab-zhwi.4.1)
